+++ app/src/main/java/org/simple/clinic/recentpatient/AllRecentPatientsEvent.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.recentpatient
+
+import org.simple.clinic.widgets.UiEvent
+
+sealed class AllRecentPatientsEvent: UiEvent