+++ app/src/main/java/org/simple/clinic/recentpatient/AllRecentPatientsUiRenderer.kt
@@ -1,0 +1,11 @@
+package org.simple.clinic.recentpatient
+
+import org.simple.clinic.mobius.ViewRenderer
+
+class AllRecentPatientsUiRenderer(
+    private val ui: AllRecentPatientsUi
+) : ViewRenderer<AllRecentPatientsModel> {
+
+  override fun render(model: AllRecentPatientsModel) {
+  }
+}