+++ app/src/main/java/org/simple/clinic/recentpatient/RecentPatientsScreenController.kt
@@ -5,7 +5,6 @@ import io.reactivex.ObservableSource
 import io.reactivex.ObservableTransformer
 import io.reactivex.rxkotlin.ofType
 import org.simple.clinic.ReplayUntilScreenIsDestroyed
-import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.patient.DateOfBirth
 import org.simple.clinic.patient.PatientRepository
@@ -21,7 +20,7 @@ import java.time.format.DateTimeFormatter
 import javax.inject.Inject
 import javax.inject.Named
 
-typealias Ui = RecentPatientsScreen
+typealias Ui = AllRecentPatientsUi
 typealias UiChange = (Ui) -> Unit
 
 class RecentPatientsScreenController @Inject constructor(
@@ -34,7 +33,6 @@ class RecentPatientsScreenController @Inject constructor(
 
   override fun apply(events: Observable<UiEvent>): ObservableSource<UiChange> {
     val replayedEvents = ReplayUntilScreenIsDestroyed(events)
-        .compose(ReportAnalyticsEvents())
         .replay()
 
     return Observable.mergeArray(